example/requirements.txt
numpy
pandas
ipykernel
tqdm
binance-connector
torch
matplotlib
deepmerge
